added login function and created valid users
user class is the creation of a user who will be apart of the apartment books program. This user can be of only one role and has different functionalities based on what the role is specifically.  Each user is assigned a apartment building to keep vital information close to each other

valid users class is list of all created user objects and has the functionalities of adding and validating users in the program